Added a new folder titled "UserOptions" containing 4 new files: - Options.h - Options.cpp - DisplayOptions.h - DisplayOptions.cpp - The Options header and source files define all the enums to be used with switch/case statements.
The DisplayOptions header and source files handle displaying the different options to the user.

Added a new file, PromptOptions.h and PromptOptions.cpp, in the UserInput folder. This file contains a function responsible for prompting the user to choose from the options displayed by the display functions.

Additionally, the previous files in the UserInput folder (Prompt.h and Prompt.cpp) were renamed to PromptAttribute to avoid confusion with the new PromptOptions files.

In the ProductManagement folder, added 4 new files:
- ProductSearch.h
- ProductSearch.cpp
- ProductDisplay.h
- ProductDisplay.cpp

The ProductSearch files contain three functionalities:
1) A function to retrieve a search predicate
2) A function to find products that match the search predicate
3) A function to orchestrate the previous two

The ProductDisplay files perform the same role as the viewAllProducts function, however, it performs a check to see whether any products were found.